# [xformers] Force to use xformers kernels